+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- Support of new transaction message format added in Exonum 0.10. (#534)
 
 ### Changed
 - `com.exonum.binding.storage.indices.MapEntry` moved to package